updated some logic of vectorinfo
